+++ oshi-core/src/main/java/oshi/util/ParseUtil.java
@@ -460,7 +460,7 @@ public class ParseUtil {
             }
             milliseconds += parseLongOrDefault(m.group(3), 0L) * 60000L;
             milliseconds += parseLongOrDefault(m.group(4), 0L) * 1000L;
-            milliseconds += 1000 * parseDoubleOrDefault("0." + m.group(5), 0d);
+            milliseconds += (long) (1000 * parseDoubleOrDefault("0." + m.group(5), 0d));
             return milliseconds;
         }
         return defaultLong;
